automation/requirements.txt
requests
beautifulsoup4
lxml
schedule
sqlalchemy
psycopg2-binary
# transformers
# torch
